Move IsNearby method to Subscription class.
