+++ src/Services/Subscriptions/SubscriptionProcessorService.cs
@@ -168,7 +168,7 @@
                         }
 
                         // Skip if not nearby or within set global location, individual subscription locations, or geofence does not match
-                        if (!IsNearby(user, pkmnCoord, true, pkmnSub.Location, pkmnSub.Areas, geofence.Name.ToLower()))
+                        if (!user.IsNearby(pkmnCoord, true, pkmnSub.Location, pkmnSub.Areas, geofence.Name.ToLower()))
                             continue;
 
                         var embed = await pokemon.GenerateEmbedMessageAsync(new AlarmMessageSettings
@@ -322,7 +322,7 @@
                         }
 
                         // Skip if not nearby or within set global location, individual subscription locations, or geofence does not match
-                        if (!IsNearby(user, pkmnCoord, true, pkmnSub.Location, pkmnSub.Areas, geofence.Name.ToLower()))
+                        if (!user.IsNearby(pkmnCoord, true, pkmnSub.Location, pkmnSub.Areas, geofence.Name.ToLower()))
                             continue;
 
                         var embed = await pokemon.GenerateEmbedMessageAsync(new AlarmMessageSettings
@@ -439,7 +439,7 @@
                         }
 
                         // Skip if not nearby or within set global location, individual subscription locations, or geofence does not match
-                        if (!IsNearby(user, raidCoord, true, subRaid.Location, subRaid.Areas, geofence.Name.ToLower()))
+                        if (!user.IsNearby(raidCoord, true, subRaid.Location, subRaid.Areas, geofence.Name.ToLower()))
                             continue;
 
                         var embed = await raid.GenerateEmbedMessageAsync(new AlarmMessageSettings
@@ -552,7 +552,7 @@
                     var geofenceMatches = questSub.Areas.Select(x => x.ToLower()).Contains(geofence.Name.ToLower());
 
                     // Skip if not nearby or within set global location, individual subscription locations, or geofence does not match
-                    if (!IsNearby(user, questCoord, true, questSub.Location, questSub.Areas, geofence.Name.ToLower()))
+                    if (!user.IsNearby(questCoord, true, questSub.Location, questSub.Areas, geofence.Name.ToLower()))
                         continue;
 
                     var embed = await quest.GenerateEmbedMessageAsync(new AlarmMessageSettings
@@ -672,7 +672,7 @@
                     }
 
                     // Skip if not nearby or within set global location, individual subscription locations, or geofence does not match
-                    if (!IsNearby(user, invasionCoord, true, invasionSub.Location, invasionSub.Areas, geofence.Name.ToLower()))
+                    if (!user.IsNearby(invasionCoord, true, invasionSub.Location, invasionSub.Areas, geofence.Name.ToLower()))
                         continue;
 
                     var embed = await pokestop.GenerateEmbedMessageAsync(new AlarmMessageSettings
@@ -778,7 +778,7 @@
                     }
 
                     // Skip if not nearby or within set global location, individual subscription locations, or geofence does not match
-                    if (!IsNearby(user, lureCoord, true, lureSub.Location, lureSub.Areas, geofence.Name.ToLower()))
+                    if (!user.IsNearby(lureCoord, true, lureSub.Location, lureSub.Areas, geofence.Name.ToLower()))
                         continue;
 
                     var embed = await pokestop.GenerateEmbedMessageAsync(new AlarmMessageSettings
@@ -893,7 +893,7 @@
                     }
 
                     // Skip if not nearby or within set global location or individual subscription locations
-                    if (!IsNearby(user, gymCoord, checkGeofence: false))
+                    if (!user.IsNearby(gymCoord, checkGeofence: false))
                         continue;
 
                     var embed = await raid.GenerateEmbedMessageAsync(new AlarmMessageSettings
@@ -1119,28 +1119,6 @@
                 maxRange = atk;
             }
             return (minRange, maxRange);
-        }
-
-        // TODO: ISubscriptionLocation (string location, List<string> areas)
-        private static bool IsNearby(Subscription user, Coordinate coord, bool checkGeofence = false, string webhookLocationName = null, List<string> areas = null, string geofenceName = null)
-        {
-            var globalLocation = user.Locations?.FirstOrDefault(x => string.Compare(x.Name, user.Location, true) == 0);
-            var webhookLocation = user.Locations?.FirstOrDefault(x => string.Compare(x.Name, webhookLocationName, true) == 0);
-            var globalDistanceMatches = globalLocation?.DistanceM > 0
-                && globalLocation?.DistanceM > new Coordinate(globalLocation?.Latitude ?? 0, globalLocation?.Longitude ?? 0).DistanceTo(coord);
-            var webhookDistanceMatches = webhookLocation?.DistanceM > 0
-                && webhookLocation?.DistanceM > new Coordinate(webhookLocation?.Latitude ?? 0, webhookLocation?.Longitude ?? 0).DistanceTo(coord);
-
-            // Skip if set distance does not match and no geofences match...
-            var matchesLocation = globalDistanceMatches || webhookDistanceMatches;
-            if (checkGeofence)
-            {
-                var geofenceNameLower = geofenceName.ToLower();
-                var matchesGeofence = areas?.Select(x => x.ToLower())
-                                            .Contains(geofenceNameLower) ?? false;
-                return matchesGeofence || matchesLocation;
-            }
-            return matchesLocation;
         }
 
         private static bool DiscordExists(ulong guildId, Dictionary<ulong, DiscordServerConfig> servers, IReadOnlyDictionary<ulong, DiscordClient> discordClients)
